+++ packages/eslint-config-airbnb-base/rules/style.js
@@ -2,6 +2,9 @@ module.exports = {
   'rules': {
     // enforce spacing inside array brackets
     'array-bracket-spacing': [2, 'never'],
+    // enforce spacing inside single-line blocks
+    // http://eslint.org/docs/rules/block-spacing
+    'block-spacing': [2, 'always'],
     // enforce one true brace style
     'brace-style': [2, '1tbs', { 'allowSingleLine': true }],
     // require camel case names
@@ -24,7 +27,7 @@ module.exports = {
     // (variable names, property names etc.)
     'id-length': 0,
     // this option sets a specific tab width for your code
-    // https://github.com/eslint/eslint/blob/master/docs/rules/indent.md
+    // http://eslint.org/docs/rules/indent
     'indent': [2, 2, { 'SwitchCase': 1, 'VariableDeclarator': 1 }],
     // specify whether double or single quotes should be used in JSX attributes
     // http://eslint.org/docs/rules/jsx-quotes
@@ -53,12 +56,17 @@ module.exports = {
     }],
     // specify the maximum depth callbacks can be nested
     'max-nested-callbacks': 0,
+    // restrict the number of statements per line
+    // http://eslint.org/docs/rules/max-statements-per-line
+    'max-statements-per-line': [0, { 'max': 1 }],
     // require a capital letter for constructors
     'new-cap': [2, { 'newIsCap': true }],
     // disallow the omission of parentheses when invoking a constructor with no arguments
     'new-parens': 0,
     // allow/disallow an empty newline after var statement
     'newline-after-var': 0,
+    // http://eslint.org/docs/rules/newline-before-return
+    'newline-before-return': 0,
     // enforces new line after each method call in the chain to make it
     // more readable and easy to maintain
     // http://eslint.org/docs/rules/newline-per-chained-call
@@ -75,6 +83,9 @@ module.exports = {
     'no-mixed-spaces-and-tabs': 2,
     // disallow multiple empty lines and only one newline at the end
     'no-multiple-empty-lines': [2, { 'max': 2, 'maxEOF': 1 }],
+    // disallow negated conditions
+    // http://eslint.org/docs/rules/no-negated-condition
+    'no-negated-condition': 0,
     // disallow nested ternary expressions
     'no-nested-ternary': 2,
     // disallow use of the Object constructor
@@ -114,6 +125,9 @@ module.exports = {
     'quotes': [2, 'double', 'avoid-escape'],
     // require identifiers to match the provided regular expression
     'id-match': 0,
+    // do not require jsdoc
+    // http://eslint.org/docs/rules/require-jsdoc
+    'require-jsdoc': 0,
     // enforce spacing before and after semicolons
     'semi-spacing': [2, { 'before': false, 'after': true }],
     // require or disallow use of semicolons instead of ASI
@@ -123,7 +137,7 @@ module.exports = {
     // require or disallow space before blocks
     'space-before-blocks': 2,
     // require or disallow space before function opening parenthesis
-    // https://github.com/eslint/eslint/blob/master/docs/rules/space-before-function-paren.md
+    // http://eslint.org/docs/rules/space-before-function-paren
     'space-before-function-paren': [2, { 'anonymous': 'always', 'named': 'never' }],
     // require or disallow spaces inside parentheses
     'space-in-parens': [2, 'never'],