+++ packages/eslint-config-airbnb/base.js
@@ -1,9 +1,4 @@
 module.exports = {
-  "extends": [
-    "eslint-config-quri/legacy",
-    "eslint-config-quri/rules/es6",
-    "eslint-config-quri/rules/babel",
-    "eslint-config-quri/rules/require-path-exists"
-  ].map(require.resolve),
-  "rules": {}
+  extends: ['eslint-config-airbnb-base'].map(require.resolve),
+  rules: {},
 };