+++ packages/eslint-config-airbnb/index.js
@@ -2,7 +2,7 @@ module.exports = {
   "extends": [
     "eslint-config-quri/base",
     "eslint-config-quri/rules/strict",
-    "eslint-config-quri/rules/react",
+    './rules/react',
   ].map(require.resolve),
   rules: {}
 };