+++ packages/eslint-config-airbnb/legacy.js
@@ -1,14 +1,4 @@
 module.exports = {
-  extends: [
-    'eslint-config-quri/rules/best-practices',
-    'eslint-config-quri/rules/errors',
-    'eslint-config-quri/rules/legacy',
-    'eslint-config-quri/rules/node',
-    'eslint-config-quri/rules/style',
-    'eslint-config-quri/rules/variables'
-  ].map(require.resolve),
-  env: {
-    browser: true,
-    node: true,
-  }
+  extends: ['eslint-config-airbnb-base/legacy'].map(require.resolve),
+  rules: {},
 };